backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "claude-agents-backend"
version = "1.0.0"
description = "FastAPI backend for Claude Agents UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-frontmatter>=1.1.0",
    "PyYAML>=6.0.2",
    "aiofiles>=24.1.0",
    "websockets>=13.0",
    "watchdog>=5.0.0",
    "gitpython>=3.1.44",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "ptyprocess>=0.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
